Fix bug in unit_merge
Bug in unit_merge caused error when spatial overlap was 0 (no units overlapped). I fixed the function to only attempt merge if there is some overlap between units.

One issue with this is that the unit_id index is no longer reset, which makes some visualizations look weird, but doesn't cause functional issues with the data.